+++ R/description.R
@@ -11,23 +11,22 @@
 #'
 #' usethis consults the following sources, in this order, to set `DESCRIPTION`
 #' fields:
-#' * `fields` argument of [create_package()] or [use_description()].
-#' * `getOption("usethis.description")` or `getOption("devtools.desc")`. The
-#' devtools option is consulted only for backwards compatibility and it's
-#' recommended to switch to an option named "usethis.description".
-#' * Defaults built into usethis.
+#' * `fields` argument of [create_package()] or [use_description()]
+#' * `getOption("usethis.description")`
+#' * Defaults built into usethis
 #'
 #' The fields discovered via options or the usethis package can be viewed with
 #' `use_description_defaults()`.
 #'
 #' If you create a lot of packages, consider storing personalized defaults as a
 #' named list in an option named `"usethis.description"`. Here's an example of
-#' code to include in `.Rprofile`:
+#' code to include in `.Rprofile`, which can be opened via [edit_r_profile()]:
 #'
 #' ```
 #' options(
 #'   usethis.description = list(
-#'     `Authors@R` = 'person("Jane", "Doe", email = "jane@example.com", role = c("aut", "cre"),
+#'     `Authors@R` = 'person("Jane", "Doe", email = "jane@example.com",
+#'                           role = c("aut", "cre"),
 #'                           comment = c(ORCID = "YOUR-ORCID-ID"))',
 #'     License = "MIT + file LICENSE",
 #'     Language =  "es"
@@ -35,14 +34,18 @@
 #' )
 #' ```
 #'
+#' Prior to usethis v2.0.0, `getOption("devtools.desc")` was consulted for
+#' backwards compatibility, but now only the `"usethis.description"` option is
+#' supported.
+#'
 #' @param fields A named list of fields to add to `DESCRIPTION`, potentially
 #'   overriding default values. See [use_description()] for how you can set
 #'   personalized defaults using package options
 #' @param check_name Whether to check if the name is valid for CRAN and throw an
 #'   error if not
-#' @param roxygen If `TRUE`, sets `RoxygenNote` to current roxygen2 version.
-#' @seealso The [description chapter](https://r-pkgs.org/description.html#dependencies)
-#'   of [R Packages](https://r-pkgs.org).
+#' @param roxygen If `TRUE`, sets `RoxygenNote` to current roxygen2 version
+#' @seealso The [description chapter](https://r-pkgs.org/description.html)
+#'   of [R Packages](https://r-pkgs.org)
 #' @export
 #' @examples
 #' \dontrun{
@@ -52,7 +55,9 @@
 #'
 #' use_description_defaults()
 #' }
-use_description <- function(fields = list(), check_name = TRUE, roxygen = TRUE) {
+use_description <- function(fields = list(),
+                            check_name = TRUE,
+                            roxygen = TRUE) {
   name <- project_name()
   if (check_name) {
     check_package_name(name)
@@ -62,6 +67,7 @@ use_description <- function(fields = list(), check_name = TRUE, roxygen = TRUE) 
   lines <- desc$str(by_field = TRUE, normalize = FALSE, mode = "file")
 
   write_over(proj_path("DESCRIPTION"), lines)
+  # TODO: why am I not using a ui_*() function here?
   if (!getOption("usethis.quiet", default = FALSE)) {
     print(desc)
   }
@@ -70,7 +76,9 @@ use_description <- function(fields = list(), check_name = TRUE, roxygen = TRUE) 
 #' @rdname use_description
 #' @param package Package name
 #' @export
-use_description_defaults <- function(package = NULL, roxygen = TRUE, fields = list()) {
+use_description_defaults <- function(package = NULL,
+                                     roxygen = TRUE,
+                                     fields = list()) {
   fields <- fields %||% list()
   check_is_named_list(fields)
 
@@ -95,8 +103,7 @@ use_description_defaults <- function(package = NULL, roxygen = TRUE, fields = li
     usethis$RoxygenNote <- roxygen_note
   }
 
-  options <- getOption("usethis.description") %||%
-    getOption("devtools.desc") %||% list()
+  options <- getOption("usethis.description") %||% list()
 
   # A `person` object in Authors@R is not patched in by modifyList()
   modify_this <- function(orig, patch) {